666
1

+++ Unity/Assets/Scripts/Codes/Model/Server/Game/Account/TokenComponent.cs
@@ -5,6 +5,6 @@ namespace ET
     [ComponentOf]
     public class TokenComponent : Entity,IAwake,IDestroy
     {
-        public Dictionary<long, string> TokenDictionary = new();
+        public Dictionary<int, string> TokenDictionary = new();
     }
 }
